A quick and dirty test of passing difference of distance to the kalman estimator as opposed to absolute rx/tx times for messages.
This works in the Bitcraze lab! There are instabilities the makes the kalman estimator go loco from time to time, but when stable it delivers a reasonable position.
The advantage of this approach is that is simplifies the code both in the kalman estimator as well as in lpsTdoaTag.c.

This is a spike. It is ugly and unit tests are not updated.